# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.ExerciseCreateRequest:
    properties:
      auto_translate:
        example: true
        type: boolean
      is_bodyweight:
        example: false
        type: boolean
      is_time_based:
        example: false
        type: boolean
      muscle_group_id:
        example: 3
        type: integer
      name:
        example: Bench Press
        maxLength: 50
        type: string
    required:
    - name
    type: object
  dto.ExerciseResponse:
    properties:
      id:
        example: 12
        type: integer
      is_bodyweight:
        example: false
        type: boolean
      is_time_based:
        example: false
        type: boolean
      muscle_group:
        $ref: '#/definitions/dto.MuscleGroupResponse'
      muscle_group_id:
        example: 3
        type: integer
      name:
        example: Bench Press
        type: string
      slug:
        example: bench-press
        type: string
    type: object
  dto.ExerciseUpdateRequest:
    properties:
      is_bodyweight:
        example: false
        type: boolean
      is_time_based:
        example: false
        type: boolean
      muscle_group_id:
        example: 4
        type: integer
      name:
        example: Incline Bench Press
        maxLength: 50
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  dto.MuscleGroupCreateRequest:
    properties:
      auto_translate:
        example: true
        type: boolean
      name:
        example: Chest
        maxLength: 50
        type: string
    required:
    - name
    type: object
  dto.MuscleGroupResponse:
    properties:
      id:
        example: 3
        type: integer
      name:
        example: Chest
        type: string
      slug:
        example: chest
        type: string
    type: object
  dto.MuscleGroupUpdateRequest:
    properties:
      name:
        example: Upper Chest
        maxLength: 50
        type: string
    type: object
...
info:
  contact:
    email: dev@example.com
    name: API Support
    url: https://example.com/support
  description: Public HTTP API for the Fitness Tracker app.
  title: Fitness Tracker API
  version: "1.0"
paths:
...
  /exercises:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ExerciseResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: List exercises
      tags:
      - exercises
    post:
      consumes:
      - application/json
      parameters:
      - description: Exercise payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ExerciseCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ExerciseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Create exercise (admin)
      tags:
      - exercises
  /exercises/{id}:
    delete:
      parameters:
      - description: Exercise ID
        example: 12
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Delete exercise (admin)
      tags:
      - exercises
    get:
      parameters:
      - description: Exercise ID
        example: 12
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ExerciseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Get exercise by ID
      tags:
      - exercises
    patch:
      consumes:
      - application/json
      parameters:
      - description: Exercise ID
        example: 12
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ExerciseUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ExerciseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Update exercise (admin)
      tags:
      - exercises
...
  /muscle-groups:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MuscleGroupResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: List muscle groups
      tags:
      - muscle-groups
    post:
      consumes:
      - application/json
      parameters:
      - description: Muscle group payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MuscleGroupCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MuscleGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Create muscle group (admin)
      tags:
      - muscle-groups
  /muscle-groups/{id}:
    delete:
      parameters:
      - description: Muscle Group ID
        example: 3
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Delete muscle group (admin)
      tags:
      - muscle-groups
    get:
      parameters:
      - description: Muscle Group ID
        example: 3
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MuscleGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Get muscle group by ID
      tags:
      - muscle-groups
    patch:
      consumes:
      - application/json
      parameters:
      - description: Muscle Group ID
        example: 3
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MuscleGroupUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MuscleGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Update muscle group (admin)
      tags:
      - muscle-groups
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"